refactor(e2e): Rename shared e2e files
